ImageProcessor/requirements.txt
celery==5.3.4
Django==4.2.10
django-celery-results==2.5.1
dj-database-url==2.1.0
django-extensions==3.1.1
django-redis==5.4.0
djangorestframework==3.14.0
drf-extra-fields==3.6.1
ipdb==0.13.13
ipython==8.12.0
psycopg2==2.9.9
redis==5.0.0
pillow==10.2.0
websockets==10.4
pika===1.3.2
